remove "dirty" variables from integration set in RooBTagDecay::getCoefAnalyticalIntegral
RooBTagDecay::getCoefAnalyticalIntegral checks if its tagging parameters
depend on the integration variables. Before, it would return 0 if this
happens, which advertises that the analytical integration cannot be
performed at all. Now, it just removes the "dirty" variables from the
integration set and advertises that it can integrate over other variables.