created new structure for LHCb repository
A new file/directory structure was created to use P2VV both stand-alone and as
a package in the LHCb software framework. The root directory has the structure
of an LHCb package, with the usual "cmt", "doc", "P2VV", "src" and
"python/P2VV" subdirectories. The stand-alone version lives in the "standalone"
directory.

"standalone" has a Makefile and setup scripts. The c++ and python code files
are symlinks to their LHCb versions by default. If differences between the
stand-alone and LHCb versions are required, symlinks can be replaced by real
files.