use single decay-time acceptance functions if PDF is simultaneous in HLT1 category
If the PDF is simultaneous for the HLT1 unbiased and HLT1
exclusively biased samples, the PDFs within the RooSimultaneous
cannot depend on the HLT1 category. Multiply each of the sub-PDFs
with either the "biased" or "unbiased" acceptance function in this
case and do not use the class that can hold two acceptance
functions.